+++ Packages/com.thenathannator.plasticband/PlasticBand/Devices/FiveFretGuitar/FiveFretGuitar.cs
@@ -1,10 +1,23 @@
+using System;
 using System.Collections.Generic;
 using UnityEngine.InputSystem;
 using UnityEngine.InputSystem.Controls;
 using UnityEngine.InputSystem.Layouts;
+using UnityEngine.InputSystem.LowLevel;
 
 namespace PlasticBand.Devices
 {
+    [Flags]
+    public enum FiveFret
+    {
+        None = 0,
+        Green = 0x01,
+        Red = 0x02,
+        Yellow = 0x04,
+        Blue = 0x08,
+        Orange = 0x10
+    }
+
     /// <summary>
     /// A 5-fret guitar controller.
     /// </summary>
@@ -105,6 +118,72 @@ namespace PlasticBand.Devices
         /// </summary>
         [InputControl(displayName = "Select")]
         public ButtonControl selectButton { get; private set; }
+
+        /// <summary>
+        /// The number of frets available on the guitar.
+        /// </summary>
+        public const int FretCount = 5;
+
+        /// <summary>
+        /// Retrieves a fret control by index.<br/>
+        /// 0 = green, 4 = orange.
+        /// </summary>
+        public ButtonControl GetFret(int index)
+        {
+            switch (index)
+            {
+                case 0: return greenFret;
+                case 1: return redFret;
+                case 2: return yellowFret;
+                case 3: return blueFret;
+                case 4: return orangeFret;
+                default: throw new ArgumentOutOfRangeException(nameof(index));
+            }
+        }
+
+        /// <summary>
+        /// Retrieves a fret control by enum value.
+        /// </summary>
+        public ButtonControl GetFret(FiveFret fret)
+        {
+            switch (fret)
+            {
+                case FiveFret.Green: return greenFret;
+                case FiveFret.Red: return redFret;
+                case FiveFret.Yellow: return yellowFret;
+                case FiveFret.Blue: return blueFret;
+                case FiveFret.Orange: return orangeFret;
+                default: throw new ArgumentException($"Could not determine the fret to retrieve! Value: '{fret}'", nameof(fret));
+            }
+        }
+
+        /// <summary>
+        /// Retrives a bitmask of the current fret states.
+        /// </summary>
+        public FiveFret GetFretMask()
+        {
+            var mask = FiveFret.None;
+            if (greenFret.isPressed) mask |= FiveFret.Green;
+            if (redFret.isPressed) mask |= FiveFret.Red;
+            if (yellowFret.isPressed) mask |= FiveFret.Yellow;
+            if (blueFret.isPressed) mask |= FiveFret.Blue;
+            if (orangeFret.isPressed) mask |= FiveFret.Orange;
+            return mask;
+        }
+
+        /// <summary>
+        /// Retrives a bitmask of the fret states in the given state event.
+        /// </summary>
+        public FiveFret GetFretMask(InputEventPtr eventPtr)
+        {
+            var mask = FiveFret.None;
+            if (greenFret.IsPressedInEvent(eventPtr)) mask |= FiveFret.Green;
+            if (redFret.IsPressedInEvent(eventPtr)) mask |= FiveFret.Red;
+            if (yellowFret.IsPressedInEvent(eventPtr)) mask |= FiveFret.Yellow;
+            if (blueFret.IsPressedInEvent(eventPtr)) mask |= FiveFret.Blue;
+            if (orangeFret.IsPressedInEvent(eventPtr)) mask |= FiveFret.Orange;
+            return mask;
+        }
 
         /// <summary>
         /// Finishes setup of the device.